feat:#60 - add JWT login page and protected routes
