+++ src/auth/authContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from "react";
+
+export interface AuthContextType {
+    token: string | null;
+    login: (token: string) => void;
+    logout: () => void;
+}
+
+export const AuthContext = createContext<AuthContextType | null>(null);
