+++ src/pages/LoginPage.tsx
@@ -1,7 +1,7 @@
 import { useState, type FormEvent } from "react";
 import { useNavigate } from "react-router-dom";
 import { Map } from "lucide-react";
-import { useAuth } from "../auth/AuthContext";
+import { useAuth } from "../auth/useAuth";
 import "./LoginPage.css";
 
 const BACKEND_URL = "http://localhost:8080";
